Move folderSchemaSpec to a new file
